spots/guesses/games are all top level routes
The straw that broke the camel's back was ember convention. I can't tell
yet if I'm making this change just because of ember, or if it makes
sense.

I do fear having all the api's nested in eachother is unsustainable...
and "weird".

What we miss here is 404 as a way of authorizing that we're e.g.
creating spots for a non-existant game. Time will tell...